+++ src/components/calcite-split-button/calcite-split-button.tsx
@@ -47,6 +47,9 @@ export class CalciteSplitButton {
   /** optionally pass an icon to display at the end of the primary button - accepts Calcite UI icon names  */
   @Prop({ reflect: true }) primaryIconEnd?: string;
 
+  /** mirror the icon for the primary button  */
+  @Prop({ reflect: true }) primaryIconMirror?: "both" | "start" | "end";
+
   /** optionally pass an aria-label for the primary button */
   @Prop({ reflect: true }) primaryLabel?: string;
 
@@ -108,6 +111,7 @@ export class CalciteSplitButton {
             loading={this.loading}
             icon-start={this.primaryIconStart ? this.primaryIconStart : null}
             icon-end={this.primaryIconEnd ? this.primaryIconEnd : null}
+            icon-mirror={this.primaryIconMirror ? this.primaryIconMirror : null}
             disabled={this.disabled}
             theme={this.theme}
             onClick={this.calciteSplitButtonPrimaryClickHandler}